Rename MultipleAddInProjectX to SharedOutDirProjectX
