+++ Source/Tests/ExcelDna.AddIn.Tasks.IntegrationTests.TestTarget/ExcelDna.AddIn.Tasks.IntegrationTests.TestTarget.sln
@@ -19,6 +19,10 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SingleDnaFileNoConfigDefaul
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PreAndPostBuildEvents", "PreAndPostBuildEvents\PreAndPostBuildEvents.csproj", "{01805002-4F09-432A-B1E2-803C273671CB}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedOutDirProjectOne", "SharedOutDirProjectOne\SharedOutDirProjectOne.csproj", "{4CE13234-A53E-4539-9AFF-D03F7DF2721C}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "SharedOutDirProjectTwo", "SharedOutDirProjectTwo\SharedOutDirProjectTwo.csproj", "{C174238C-B87F-41D3-BCAA-325AB66F0A7F}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -57,6 +61,14 @@ Global
 		{01805002-4F09-432A-B1E2-803C273671CB}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{01805002-4F09-432A-B1E2-803C273671CB}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{01805002-4F09-432A-B1E2-803C273671CB}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4CE13234-A53E-4539-9AFF-D03F7DF2721C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CE13234-A53E-4539-9AFF-D03F7DF2721C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CE13234-A53E-4539-9AFF-D03F7DF2721C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CE13234-A53E-4539-9AFF-D03F7DF2721C}.Release|Any CPU.Build.0 = Release|Any CPU
+		{C174238C-B87F-41D3-BCAA-325AB66F0A7F}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{C174238C-B87F-41D3-BCAA-325AB66F0A7F}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{C174238C-B87F-41D3-BCAA-325AB66F0A7F}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{C174238C-B87F-41D3-BCAA-325AB66F0A7F}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
