+++ Source/Tests/ExcelDna.AddIn.Tasks.IntegrationTests.TestTarget/SharedOutDirProjectOne/Properties/AssemblyInfo.cs
