+++ Source/Tests/ExcelDna.AddIn.Tasks.IntegrationTests.TestTarget/SharedOutDirProjectTwo/Properties/AssemblyInfo.cs
